Summary
On app launch, Activity A page will appear on the screen with a random generated number
The Activity A and B will contain a Button and TextView widgets
The TextView widget will display a random number
The Button widget is wired with an onClick event.
When Activity A button is clicked, it will navigate to Activity B and display the generated random number on Activity B TextView widget.
When the Button widget on Activity B is clicked, it will generate a new random numbers and display it on the TextView on Activity A
The Activity A button text is “Activity A Button“ while Activity B is “Activity B Button“.